Move test to appropriate test class
